+++ Kamienica/src/main/java/kamienica/feature/payment/PaymentWaterDAOImpl.java
@@ -27,7 +27,7 @@ public class PaymentWaterDAOImpl extends AbstractDao<Integer, PaymentWater>
 	@SuppressWarnings("unchecked")
 	public List<PaymentWater> getByInvoice(Invoice invoice) {
 		Query query = getSession().createSQLQuery("Select * from paymentwater where invoice_id = :id")
-				.addEntity(PaymentWater.class).setInteger("id", invoice.getId());
+				.addEntity(PaymentWater.class).setLong("id", invoice.getId());
 		List<PaymentWater> result = query.list();
 		return result;
 	}
